[MOD] Move modals to its own folder
